✨ adds split reaminder handler
